src/dev_requirements.txt
# Project dependencies
numpy==1.26.0
matplotlib==3.8.0 
scipy==1.11.3
opencv-python==4.8.1.78
scikit-learn==1.3.1
fake-bpy-module
pyyaml==6.0.1 
tqdm==4.66.6
